+++ src/docs/CHANGELOG.md
@@ -88,6 +88,7 @@ Features:
 - Removed Errors module.
 - Removed support old function based controllers and APIs.
 - Removed events: systemerror, setup.errorreporting, frontcontroller.exception event.
+- Development mode is now controlled by editing app/config/kernel.yml `kernel = dev` or `kernel = prod`
 
 
 CHANGELOG - ZIKULA 1.3.5